+++ docs/md/releasenotes.md
@@ -31,6 +31,7 @@ make its usage easier. This includes:
 - Remove `JvmTaintMemoryLocationBamCpaRun` in order to simplify the utility code to run the analysis. `TaintAnalyzer` is the direct replacement, this class has the same behavior as `JvmTaintMemoryLocationBamCpaRun`, as long as `analyze` is called only once.
 - Remove the `CpaRun` infrastructure.
 - Remove `MapAbstractStateFactory`, `WrapperTransferRelation`, and other classes that were only used with the tree heap model.
+- Remove `StateNames` and `getStateByName`.
 
 ## Version 9.1.6